Refactored Quiz class.
